# src/main/resources/api/api.yml
openapi: 3.0.0
info:
    description: "Swagger API rest service"
    version: "1.0.0"
    title: "Swagger backend"
paths:
...
    /business:
        get:
            tags:
                - "business"
            summary: "Get data about business of the current user"
            description: "Get data about business of the current user"
            operationId: "getCurrentUserBusiness"
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessResponse'
                '400':
                    description: "Bad request"
                '404':
                    description: "Not found"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"

        post:
            tags:
                - "business"
            summary: "Add a new business"
            description: "Add a new business"
            operationId: "addBusiness"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/businessPayload'
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /get_businesses_list/{city}:
        get:
            tags:
                - "business"
            summary: "Get list of all businesses"
            description: "Get list of all businesses"
            operationId: "getBusinessesList"
            parameters:
                -   in: path
                    name: city
                    required: true
                    schema:
                        type: string
                    description: Name of the city you look for businesses
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessListResponse'
                '400':
                    description: "Bad request"
                '404':
                    description: "Not found"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /get_nearest_restaurants_list:
        post:
            tags:
                - "business"
            summary: "Get list of nearest restaurants"
            description: "Get list of nearest restaurants"
            operationId: "getNearestRestaurants"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/userLocationPayload'
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/nearestBusinessResponseList'
                '400':
                    description: "Bad request"
                '404':
                    description: "Not found"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /search_business/{name}/{longitude}/{latitude}:
        get:
            tags:
                - "business"
            summary: "Get list of businesses with given name"
            description: "Get list of businesses with given name"
            operationId: "searchBusiness"
            parameters:
                -   name: name
                    in: path
                    required: true
                    schema:
                        type: string
                        minLength: 0
                        maxLength: 50
                -   name: longitude
                    in: path
                    required: true
                    schema:
                        type: number
                        format: double
                        minimum: -180
                        maximum: 180

                -   name: latitude
                    in: path
                    required: true
                    schema:
                        type: number
                        format: double
                        minimum: -90
                        maximum: 90
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessListResponse'
                '400':
                    description: "Bad request"
                '404':
                    description: "Not found"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /find_business_by_name/{name}:
        get:
            tags:
                - "business"
            summary: "Get list of businesses with given name"
            description: "Get list of businesses with given name"
            operationId: "findBusinesses"
            parameters:
                -   in: path
                    name: name
                    required: true
                    schema:
                        type: string
                    description: Name of business you are looking for
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessListResponse'
                '400':
                    description: "Bad request"
                '404':
                    description: "Not found"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /business/top_rated:
        get:
            tags:
                - "business"
            summary: "Get sorted list of the top rated businesses"
            description: "Get sorted list of the top rated businesses"
            operationId: "getTopRated"
            parameters:
                -   name: quantity
                    in: query
                    schema:
                        type: integer
                        minimum: 1
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/businessListResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/add_product:
        post:
            tags:
                - "product"
            summary: "Add new product"
            description: "Add new product"
            operationId: "addProduct"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/productPayload'
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/productResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/get_product/{id}:
        get:
            tags:
                - "product"
            summary: "Get product from database"
            description: "Get product from database"
            operationId: "getProduct"
            parameters:
                -   in: path
                    name: id
                    required: true
                    schema:
                        type: integer
                        format: int64
                        minimum: 1
                    description: The product ID
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/productResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/search_product/{name}/{longitude}/{latitude}:
        get:
            tags:
                - "product"
            summary: "Search for product"
            description: "Search for product"
            operationId: "searchProduct"
            parameters:
                -   name: name
                    in: path
                    required: true
                    schema:
                        type: string
                        minLength: 0
                        maxLength: 50
                -   name: longitude
                    in: path
                    required: true
                    schema:
                        type: number
                        format: double
                        minimum: -180
                        maximum: 180

                -   name: latitude
                    in: path
                    required: true
                    schema:
                        type: number
                        format: double
                        minimum: -90
                        maximum: 90
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/productListResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/delete_product/{id}:
        delete:
            tags:
                - "product"
            summary: "Delete product from database"
            description: "Delete product from database"
            operationId: "deleteProduct"
            parameters:
                -   in: path
                    name: id
                    required: true
                    schema:
                        type: integer
                        format: int64
                        minimum: 1
                    description: The product ID
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                type: string
                                example: "product deleted"
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/get_products_list/{ownerId}:
        get:
            tags:
                - "product"
            summary: "Get product from database"
            description: "Get product from database"
            operationId: "getProductsList"
            parameters:
                -   in: path
                    name: ownerId
                    required: true
                    schema:
                        type: integer
                        format: int64
                        minimum: 1
                    description: The business ID
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/productListResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
    /product/search_latest:
        get:
            tags:
                - "product"
            summary: "Get the latest products"
            description: "Get the latest products"
            operationId: "getLatestProducts"
            parameters:
                -   name: quantity
                    in: query
                    schema:
                        type: integer
                        minimum: 1
            responses:
                '201':
                    description: 'Operation successful'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/productListResponse'
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
                '503':
                    description: "Database connection failed"
...
components:
    schemas:
...
        businessPayload:
            type: object
            properties:
                name:
                    type: string
                    minLength: 3
                    maxLength: 80
                country:
                    type: string
                    minLength: 4
                    maxLength: 80
                city:
                    type: string
                    minLength: 3
                    maxLength: 80
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 50
                    pattern: '^\d+-\d+$'
                street:
                    type: string
                    minLength: 3
                    maxLength: 80
                streetNumber:
                    type: string
                    minLength: 1
                    maxLength: 50
                email:
                    type: string
                    minLength: 5
                    maxLength: 200
                    pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
                phoneNumber:
                    type: string
                    minLength: 6
                    maxLength: 12
                businessType:
                    type: string
                    enum:
                        - SHOP
                        - RESTAURANT
                openHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
                closeHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
        businessResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                    minLength: 3
                    maxLength: 80
                country:
                    type: string
                    minLength: 4
                    maxLength: 80
                city:
                    type: string
                    minLength: 3
                    maxLength: 80
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 50
                    pattern: '^\d+-\d+$'
                street:
                    type: string
                    minLength: 3
                    maxLength: 80
                streetNumber:
                    type: string
                    minLength: 1
                    maxLength: 50
                email:
                    type: string
                    minLength: 5
                    maxLength: 200
                    pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
                phoneNumber:
                    type: string
                    minLength: 6
                    maxLength: 12
                businessType:
                    type: string
                    enum:
                        - SHOP
                        - RESTAURANT
                openHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
                closeHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
        nearestBusinessResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                    minLength: 3
                    maxLength: 80
                country:
                    type: string
                    minLength: 4
                    maxLength: 80
                city:
                    type: string
                    minLength: 3
                    maxLength: 80
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 50
                    pattern: '^\d+-\d+$'
                street:
                    type: string
                    minLength: 3
                    maxLength: 80
                streetNumber:
                    type: string
                    minLength: 1
                    maxLength: 50
                email:
                    type: string
                    minLength: 5
                    maxLength: 200
                    pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
                phoneNumber:
                    type: string
                    minLength: 6
                    maxLength: 12
                businessType:
                    type: string
                    enum:
                        - SHOP
                        - RESTAURANT
                openHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
                closeHour:
                    type: string
                    pattern: '^\d{2}:\d{2}$'
                distance:
                    type: integer
        nearestBusinessResponseList:
            type: array
            items:
                $ref: '#/components/schemas/nearestBusinessResponse'
        productPayload:
            type: object
            properties:
                name:
                    type: string
                    minLength: 3
                    maxLength: 80
                regularPrice:
                    type: number
                    format: double
                    maximum: 1000
                discountedPrice:
                    type: number
                    format: double
                    maximum: 1000
                expirationDate:
                    type: string
                    pattern: '^\d{4}-\d{2}-\d{2}$'
                    example: "2019-05-17"
                vegan:
                    type: boolean
                image:
                    type: string
        productResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                ownerId:
                    type: number
                    maxLength: 120
                name:
                    type: string
                    minLength: 3
                    maxLength: 80
                regularPrice:
                    type: number
                    format: double
                    maximum: 1000
                discountedPrice:
                    type: number
                    format: double
                    maximum: 1000
                expirationDate:
                    type: string
                    pattern: '^\d{4}-\d{2}-\d{2}$'
                    example: "2019-05-17"
                vegan:
                    type: boolean
                linkToResource:
                    type: string
...
        productListResponse:
            type: array
            items:
                $ref: '#/components/schemas/productResponse'
        businessListResponse:
            type: array
            items:
                $ref: '#/components/schemas/businessResponse'
        userLocationPayload:
            type: object
            properties:
                city:
                    type: string
                    minLength: 0
                    maxLength: 80
                street:
                    type: string
                    minLength: 0
                    maxLength: 80
                streetNumber:
                    type: string
                    minLength: 0
                    maxLength: 50
                businessType:
                    type: string
                    enum:
                        - SHOP
                        - RESTAURANT
                longitude:
                    type: number
                    format: double
                    minimum: -180
                    maximum: 180
                latitude:
                    type: number
                    format: double
                    minimum: -90
                    maximum: 90